bug fix in load64

+++ src/mem.cc
@@ -3,9 +3,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include <bitset>
 #include <cassert>
 #include <cmath>
 #include <cstdint>
+#include <iostream>
 
 #include "log.h"
 
@@ -115,6 +117,7 @@ uint32_t Mem::load32(uint64_t addr) {
 uint64_t Mem::load64(uint64_t addr) {
   uint8_t *p = (uint8_t *)addr;
 
-  return p[0] | p[1] << 8 | p[2] << 16 | p[3] << 24 | uint64_t(p[4]) << 32 |
-         uint64_t(p[5]) << 40 | uint64_t(p[6]) << 48 | uint64_t(p[7]) << 56;
+  return uint64_t(p[0]) | uint64_t(p[1]) << 8 | uint64_t(p[2]) << 16 |
+         uint64_t(p[3]) << 24 | uint64_t(p[4]) << 32 | uint64_t(p[5]) << 40 |
+         uint64_t(p[6]) << 48 | uint64_t(p[7]) << 56;
 }
